requirements.txt
torch
transformers
faster-whisper
isodate
google-api-python-client
yt-dlp
grpcio
grpcio-tools
streamlit
protobuf
numpy
pandas